requirements.txt
absl-py==2.2.2
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
flatbuffers==25.2.10
fonttools==4.57.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
graphtools==1.5.3
grpcio==1.71.0
h5py==3.13.0
idna==3.10
igraph==0.10.8
imageio==2.37.0
joblib==1.5.1
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
leidenalg==0.9.1
libclang==18.1.1
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
networkx==3.5
numba==0.61.2
numpy==1.26.4
opencv-python==4.11.0.86
openslide-bin==4.0.0.6
openslide-python==1.4.1
openTSNE==1.0.1
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.0.3
phate==1.0.11
pillow==11.2.1
protobuf==5.29.4
Pygments==2.19.1
PyGSP==0.5.1
pynndescent==0.5.13
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-igraph==0.10.8
pytz==2025.2
requests==2.32.3
rich==14.0.0
s_gd2==1.8.1
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.15.3
scprep==1.2.3
six==1.17.0
staintools==2.1.2
tasklogger==1.2.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.5
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
